Pair without prefix
When pair with other nodes only send the suffix, because the prefix is
always equal to the paired nodes own prefix.
If the prefix is changed from the border router, the link will still be
valid.
If paired with a device outside the adress space, the entire ipv6
address should be send.